Release version 1.0.0

+++ src/Bullet.java
@@ -2,14 +2,14 @@ package sotasadventure;
 
 /**
  * The class that represents a bullet.
+ * @author Sota Nishiyama
  */
-
 public class Bullet {
-    public static final float DURATION = (float) .7;
-    public static final float SUCCESSION = (float) .6;
+    public static final double DURATION = .7;
+    public static final double SUCCESSION = .6;
 
     public Vector position = new Vector();
     public int speed;
     public int damage = 1;
-    public float firedTime;
+    public double firedTime;
 }
