+++ src/State.java
@@ -5,18 +5,16 @@ import javax.swing.JPanel;
 /**
  * The super class of states.
  * This class contains some methods that is used to update contents and help transition.
- * 
+ *
  * @author Sota Nishiyama
- * @since  1.0
  */
 public class State extends JPanel {
 
     /**
      * Updates contents.
-     * 
      * @param elapsedTime The time elapsed since the start of the game
      */
-    public void update(float elapsedTime) {}
+    public void update(double elapsedTime) {}
 
     /**
      * Render changes.
@@ -30,7 +28,6 @@ public class State extends JPanel {
 
     /**
      * Called at the beginning of the state.
-     * 
      * @param opt the optional variable
      */
     public void enter(String opt) {}
