+++ src/StateMachine.java
@@ -9,11 +9,9 @@ import javax.swing.JPanel;
  * The class that stores and handles all states used in the game.
  * When updating and rendering contents in the game, methods in this class will be called instead of methods in each State.
  * The class sets the current state to currentStage field and operates it through other methods.
- * 
+ *
  * @author Sota Nishiyama
- * @since  1.0
  */
-
 public class StateMachine {
 
     private HashMap<String, State> states = new HashMap<String, State>();
@@ -22,7 +20,6 @@ public class StateMachine {
 
     /**
      * Sets the panel this state machine is applied to.
-     * 
      * @param panel the panel this state machine is applied to
      */
     public StateMachine(JPanel panel) {
@@ -32,10 +29,9 @@ public class StateMachine {
 
     /**
      * Calls the update method of the current state.
-     * 
      * @param elapsedTime the time elapsed since the game started
      */
-    public void update(float elapsedTime) {
+    public void update(double elapsedTime) {
         currentState.update(elapsedTime);
     }
 
@@ -48,7 +44,6 @@ public class StateMachine {
 
     /**
      * Changes the current state.
-     * 
      * @param name the name of the state which comes next
      */
     public void change(String name) {
@@ -63,7 +58,6 @@ public class StateMachine {
 
     /**
      * Changes the current state and give a value to the next state.
-     * 
      * @param name the name of the state which comes next
      * @param opt the optional variable
      */
@@ -79,7 +73,6 @@ public class StateMachine {
 
     /**
      * Adds a state to the state machine.
-     * 
      * @param name the name of the state to be added
      * @param state the state to be added
      */
